// src/app/page/master/kategori-dokumen/kategori-dokumen.interface.ts
import { EventEmitter } from '@angular/core';

export interface KategoriDokumen {
    kode?: any;
    namaKategoryDokumen?: string;
    reportDisplay?: string;
    kodeExternal?: string;
    namaExternal?: string;
    statusEnabled?: boolean;
    noRec?: string;
    kategoryDokumen?: any;
    id?: any;

}